Support narrowed buffers
Using screen lines instead of line numbers to calculate the nlines
allows to support narrowed buffers.

Note that `count-screen-lines' is an expensive operation especially
in a large buffer. It calles `vertical-motion' with `buffer-size'.
To get the screen-lines greater than the screen-height is of no
interest here, so introduce `vim-empty-lines-count-screen-lines' and
`vim-empty-lines-setup-count-screen-lines'.

Bump Package-Requires emacs 23.2 for `with-silent-modifications'.

Signed-off-by: Takeshi Banse <takebi@laafc.net>